requirements.txt
google-cloud-speech==0.27.1
google-cloud-logging==1.1.0
tornado==4.5.1